refactor: dedicated application callback module
